Add a main Makefile and set travis to use it
Travis will now build on the overall makefile by using
`make test`. This probably won't work right now. There are a
couple of frontend tests that are failing, and I do not know if
travis will be able to run the selenium tests by default. I do
know that there is a way to set those up using travis, though.